backend/JWTtoken.py
from datetime import datetime, timedelta, timezone
import jwt


SECRET_KEY = "42c9303f1ce70ef5c8495ec89725d1b496671b4c207bb2f6b6f2035b4a21fcab"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
