backend/requirements.txt
fastapi[standard]
sqlmodel
passlib[bcrypt]
pyjwt